UIのリファクタリング: プレイリストビューを追加し、アルバムおよびアーティストビューのレイアウトを改善
